// big-practice/frontend/src/types/teacher.type.ts
export interface ITeacher {
  _id: string;
  name: string;
  email: string;
  password: string;
  phoneNumber: string;
  selectedClass: { name: string };
  selectedGender: 'Male' | 'Female' | 'Other';
  subjects: { name: string }[];
}
// _id: number;
//   name: string;
//   subject: string;
//   email: string;
//   class: string;
//   gender: 'Male' | 'Female' | 'Other';